delta overflow solved

+++ initial_tests/speedRegulator/Core/Src/encoders.c
@@ -8,11 +8,11 @@
 
 #include "encoders.h"
 
-int16_t encoderL = 0;
-int16_t encoderR = 0;
+uint16_t encoderL = 0;
+uint16_t encoderR = 0;
 
-int16_t encoderLOld = 0;
-int16_t encoderROld = 0;
+uint16_t encoderLOld = 0;
+uint16_t encoderROld = 0;
 float encoderLSpeed = 0;
 float encoderRSpeed = 0;
 
@@ -29,10 +29,10 @@ void getEncoders(TIM_HandleTypeDef *T2, TIM_HandleTypeDef *T3){
 	  encoderL = __HAL_TIM_GET_COUNTER(T2);
 	  encoderR = __HAL_TIM_GET_COUNTER(T3);
 
-	  encoderLSpeed = (float)(encoderL-encoderLOld)/encoderTd * impConst; // mm/sec
+	  encoderLSpeed = (float)((int16_t)(encoderL-encoderLOld))/encoderTd * impConst; // mm/sec
 	  encoderLOld = encoderL;
 
-	  encoderRSpeed = (float)(encoderR-encoderROld)/encoderTd * impConst; // mm/sec
+	  encoderRSpeed = (float)((int16_t)(encoderR-encoderROld))/encoderTd * impConst; // mm/sec
 	  encoderROld = encoderR;
